Separate payments datasource
